requirements.txt
certifi==2022.6.15.2
charset-normalizer==2.1.1
classproperty==1.0
docutils==0.19
EasyProcess==1.1
entrypoint2==1.1
idna==3.4
Kivy==2.1.0
kivy-deps.angle==0.3.2
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
mss==6.1.0
Pillow==9.2.0
Pygments==2.13.0
pynput==1.7.6
pyperclip==1.8.2
pypiwin32==223
pyscreenshot==3.0
pywin32==304
requests==2.28.1
six==1.16.0
urllib3==1.26.12
